classes refactoring
