+++ question/bank/managecategories/classes/question_category_list.php
@@ -1,0 +1,98 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * Class representing a list of question categories
+ *
+ * @copyright  1999 onwards Martin Dougiamas {@link http://moodle.com}
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace qbank_managecategories;
+
+defined('MOODLE_INTERNAL') || die();
+
+require_once($CFG->libdir. '/listlib.php');
+
+use stdClass;
+use moodle_list;
+
+class question_category_list extends moodle_list {
+    public $table = "question_categories";
+    public $listitemclassname = '\qbank_managecategories\question_category_list_item';
+    /**
+     * Reference to list displayed below this one.
+     * @var $nextlist
+     */
+    public $nextlist = null;
+    /**
+     * Reference to list displayed above this one.
+     * @var $lastlist
+     */
+    public $lastlist = null;
+
+    public $context = null;
+    public $sortby = 'parent, sortorder, name';
+
+    public function __construct($type='ul', $attributes='', $editable = false, $pageurl=null, $page = 0, $pageparamname = 'page', $itemsperpage = 20, $context = null){
+        parent::__construct('ul', '', $editable, $pageurl, $page, 'cpage', $itemsperpage);
+        $this->context = $context;
+    }
+
+    public function get_records() {
+        $this->records = get_categories_for_contexts($this->context->id, $this->sortby);
+    }
+
+    /**
+     * Returns the highest category id that the $item can have as its parent.
+     * Note: question categories cannot go higher than the TOP category.
+     *
+     * @param list_item $item The item which its top level parent is going to be returned.
+     * @return int
+     */
+    public function get_top_level_parent_id($item) {
+        // Put the item at the highest level it can go.
+        $topcategory = question_get_top_category($item->item->contextid, true);
+        return $topcategory->id;
+    }
+
+    /**
+     * process any actions.
+     *
+     * @param integer $left id of item to move left
+     * @param integer $right id of item to move right
+     * @param integer $moveup id of item to move up
+     * @param integer $movedown id of item to move down
+     * @return void
+     */
+    public function process_actions($left, $right, $moveup, $movedown) {
+        $category = new stdClass();
+        if (!empty($left)) {
+            // Moved Left (In to another category).
+            $category->id = $left;
+            $category->contextid = $this->context->id;
+            $event = \core\event\question_category_moved::create_from_question_category_instance($category);
+            $event->trigger();
+        } else if (!empty($right)) {
+            // Moved Right (Out of the current category).
+            $category->id = $right;
+            $category->contextid = $this->context->id;
+            $event = \core\event\question_category_moved::create_from_question_category_instance($category);
+            $event->trigger();
+        }
+        parent::process_actions($left, $right, $moveup, $movedown);
+    }
+}
